# src/openapi.yaml
openapi: '3.1.0'
info:
  title: WIZBI Control Plane API
  version: '1.0.0'
  description: |
    REST API for the WIZBI Control Plane — manages organizations, projects, 
    templates, users, and GCP/GitHub infrastructure provisioning.
    
    **Authentication**: All endpoints require either:
    - `X-Firebase-ID-Token` header (for web UI)
    - `X-API-Key` header (for programmatic/agent access)
    
    **MCP Server**: Available at `/api/mcp/sse` for AI agent integration.
  license:
    name: MIT
servers:
  - url: /api
    description: Local / deployed server
components:
  securitySchemes:
    firebaseToken:
      type: apiKey
      in: header
      name: X-Firebase-ID-Token
      description: Firebase ID Token for web UI auth
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for programmatic access
...
security:
  - firebaseToken: []
  - apiKey: []
paths:
...
  /orgs:
    get:
      tags: [Organizations]
      summary: List organizations
      responses:
        '200':
          description: List of organizations
    post:
      tags: [Organizations]
      summary: Create organization
      description: Creates a GCP folder, GitHub team, and Firestore doc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                phone: { type: string }
      responses:
        '201':
          description: Organization created
  /orgs/{id}:
    delete:
      tags: [Organizations]
      summary: Delete organization
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Deletion started

  # ─── Projects ───────────────────────────────────────
...